+++ NEWS.md
@@ -1,3 +1,7 @@
+EnsemblePatientLevelPrediction 0.0.2
+======================
+- Added stacker ensemble code (currently only for logistic regression)
+  
 EnsemblePatientLevelPrediction 0.0.1
 ======================
-  - A new package that wraps around PatientLevelPrediction to create and evaluate ensemble models
+- A new package that wraps around PatientLevelPrediction to create and evaluate ensemble models